+++ Anvil/CMakeLists.txt
@@ -6,11 +6,8 @@ set(CMAKE_CXX_STANDARD 17)
 
 add_executable(Anvil
                Anvil.m.cpp
-               WorldLayer.cpp
-
-               EscapeMenu.cpp
-
-               EditorUI.cpp)
+               EditorLayer.cpp
+               EscapeMenu.cpp)
 
 target_link_libraries(Anvil PRIVATE Sprocket)
 target_include_directories(Anvil PUBLIC .)
