+++ Anvil/EditorLayer.cpp
@@ -1,4 +1,4 @@
-#include "WorldLayer.h"
+#include "EditorLayer.h"
 
 WorldLayer::WorldLayer(const Sprocket::CoreSystems& core) 
     : Sprocket::Layer(core)
