+++ Anvil/EscapeMenu.h
@@ -1,8 +1,6 @@
 #pragma once
 #include <Sprocket.h>
-
-#include "WorldLayer.h"
-#include "EditorUI.h"
+#include "EditorLayer.h"
 
 #include <memory>
 
@@ -15,8 +13,7 @@ class EscapeMenu : public Sprocket::Layer
 
 public:
     EscapeMenu(const Sprocket::CoreSystems& core,
-               WorldLayer* worldLayer,
-               EditorUI* editorUi);
+               WorldLayer* worldLayer);
 
     void OnEvent(Sprocket::Event& event) override;
     void OnUpdate(double dt) override;
